+++ src/pashmak.py
@@ -36,6 +36,16 @@ if __name__ == '__main__':
     # set signal handler
     signal.signal(signal.SIGINT, signal_handler)
 
+    # set the python path
+    try:
+        py_path = os.environ['PYTHONPATH'].split(':')
+        for path in py_path:
+            path = path.strip()
+            if path != '':
+                sys.path.append(path)
+    except:
+        pass
+
     # validate arguments
     if len(sys.argv) <= 1:
         print(sys.argv[0] + ': script file name is required: pashmak [filename]')