+++ src/pashmak.py
@@ -36,6 +36,14 @@ if __name__ == '__main__':
     # set signal handler
     signal.signal(signal.SIGINT, signal_handler)
 
+    # set the python path
+    try:
+        py_path = os.environ['PYTHONPATH'].split(':')
+        py_path = [p for p in py_path if p != '']
+        sys.path = [*py_path, *sys.path]
+    except:
+        pass
+
     # validate arguments
     if len(sys.argv) <= 1:
         print(sys.argv[0] + ': script file name is required: pashmak [filename]')