+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - added general variable `$__ismain__`
 - added **Jit** compiler
 - added `fopen` alias for `open` function
+- added `py_load_file` function to load python scripts as object (GH-35)
 
 #### Bug Fixes
 - fixed a bug in `eval` function