+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - added **Jit** compiler
 - added `fopen` alias for `open` function
 - added `py_load_file` function to load python scripts as object (GH-35)
+- added python `json` library support
 
 #### Bug Fixes
 - fixed a bug in `eval` function